+++ lib/sequel/extensions/activerecord_connection/mysql2.rb
@@ -1,34 +1,17 @@
 module Sequel
   module ActiveRecordConnection
     module Mysql2
-      def execute(sql, opts=OPTS)
-        original_query_options = activerecord_raw_connection.query_options.dup
+      def synchronize(*)
+        super do |conn|
+          original_query_options = conn.query_options.dup
+          conn.query_options.merge!(as: :hash, symbolize_keys: true, cache_rows: false)
 
-        activerecord_raw_connection.query_options.merge!(
-          as:             :hash,
-          symbolize_keys: true,
-          cache_rows:     false,
-        )
-
-        result = activerecord_connection.execute(sql)
-
-        if opts[:type] == :select
-          if block_given?
-            yield result
-          else
-            result
+          begin
+            yield conn
+          ensure
+            conn.query_options.replace(original_query_options)
           end
-        elsif block_given?
-          yield activerecord_raw_connection
         end
-      rescue ActiveRecord::StatementInvalid => exception
-        if exception.cause.is_a?(::Mysql2::Error)
-          raise_error(exception.cause)
-        else
-          raise
-        end
-      ensure
-        activerecord_raw_connection.query_options.replace(original_query_options)
       end
     end
   end